requirements.txt
pyrogram
python-dotenv
google-generativeai
tgcrypto
